minors

+++ examples/wordcount.cc
@@ -13,41 +13,44 @@
 #include <sstream>
 #include <string>
 
-int main(int argc, const char* argv[]) {
-  klex::regular::Compiler cc;
-  cc.parse(R"(
-      Word  ::= [a-zA-Z]+
-      LF    ::= \n
-      Other ::= .
-      Eof   ::= <<EOF>>
-  )");
+int main(int argc, const char* argv[])
+{
+	klex::regular::Compiler cc;
+	cc.parse(R"(
+		Word  ::= [a-zA-Z]+
+		LF    ::= \n
+		Other ::= .
+		Eof   ::= <<EOF>>
+	)");
 
-  int words = 0;
-  int chars = 0;
-  int lines = 0;
+	size_t words = 0;
+	size_t chars = 0;
+	size_t lines = 0;
 
-  klex::regular::Lexer<int, int, false, false> lexer { cc.compile(), std::cin };
-  for (bool eof = false; !eof; ) {
-    switch (lexer.recognize()) {
-      case 4: // EOF
-        eof = true;
-        break;
-      case 3: // Other
-        assert(lexer.token() == 3);
-        chars++;
-        break;
-      case 2: // LF
-        chars++;
-        lines++;
-        break;
-      case 1: // Word
-        words++;
-        chars += lexer.word().size();
-        break;
-    }
-  }
+	klex::regular::Lexer<int, int, false, false> lexer{cc.compile(), std::cin};
+	for (bool eof = false; !eof;)
+	{
+		switch (lexer.recognize())
+		{
+			case 4:  // EOF
+				eof = true;
+				break;
+			case 3:  // Other
+				assert(lexer.token() == 3);
+				chars++;
+				break;
+			case 2:  // LF
+				chars++;
+				lines++;
+				break;
+			case 1:  // Word
+				words++;
+				chars += lexer.word().size();
+				break;
+		}
+	}
 
-  std::cout << "newlines: " << lines << ", words: " << words << ", characters: " << chars << "\n";
+	std::cout << "newlines: " << lines << ", words: " << words << ", characters: " << chars << "\n";
 
-  return EXIT_SUCCESS;
+	return EXIT_SUCCESS;
 }
